warped_mppi/mppi_warp.py
"""
Warp-based MPPI implementation.

This module contains the NVIDIA Warp implementation of the MPPI controller,
migrated from the original PyCUDA version.
"""

import warp as wp
import numpy as np
from typing import Optional, Tuple, List, Any


# Initialize Warp
wp.init()


# Warp data structures
@wp.struct
class State:
    """Vehicle state representation."""

    x: wp.float32  # Position x
    y: wp.float32  # Position y
    v: wp.float32  # Velocity
    theta: wp.float32  # Heading angle


@wp.struct
class Control:
    """Vehicle control input."""

    a: wp.float32  # Acceleration
    delta: wp.float32  # Steering angle


@wp.struct
class Obstacle:
    """Circular obstacle representation."""

    x: wp.float32  # Center x
    y: wp.float32  # Center y
    radius: wp.float32  # Obstacle radius
    min_x: wp.float32  # Minimum distance point x
    min_y: wp.float32  # Minimum distance point y
    distance: wp.float32  # Distance to vehicle


@wp.struct
class OptimizationParams:
    """MPPI optimization parameters."""

    samples: wp.int32
    M: wp.float32
    dt: wp.float32
    num_controls: wp.int32
    num_obstacles: wp.int32
    vehicle_length: wp.float32
    c_lambda: wp.float32
    scan_range: wp.float32
    steering_rate_weight: wp.float32
    method: wp.int32


@wp.struct
class CostWeights:
    """Cost function weight matrices."""

    Q: wp.vec4  # State tracking weights [x, y, v, theta]
    Qf: wp.vec4  # Final state weights
    R: wp.vec2  # Control effort weights [a, delta]


@wp.struct
class Limits:
    """Control limits and noise parameters."""

    u_limits: wp.vec2  # Maximum control values [a_max, delta_max]
    u_dist_limits: wp.vec2  # Noise standard deviations


# Mathematical utility functions
@wp.func
def euler_step(state: State, control: Control, vehicle_length: wp.float32) -> State:
    """Single Euler integration step for bicycle model dynamics."""
    result = State()
    result.x = state.v * wp.cos(state.theta)
    result.y = state.v * wp.sin(state.theta)
    result.v = control.a
    result.theta = state.v * wp.tan(control.delta) / vehicle_length
    return result


@wp.func
def update_state(state: State, derivative: State, dt: wp.float32) -> State:
    """Update state using derivative and timestep."""
    result = State()
    result.x = state.x + derivative.x * dt
    result.y = state.y + derivative.y * dt
    result.v = state.v + derivative.v * dt
    result.theta = state.theta + derivative.theta * dt
    return result


@wp.func
def runge_kutta_step(
    state: State, control: Control, vehicle_length: wp.float32, dt: wp.float32
) -> State:
    """4th-order Runge-Kutta integration step."""
    k1 = euler_step(state, control, vehicle_length)

    tmp_state = update_state(state, k1, dt * 0.5)
    k2 = euler_step(tmp_state, control, vehicle_length)

    tmp_state = update_state(state, k2, dt * 0.5)
    k3 = euler_step(tmp_state, control, vehicle_length)

    tmp_state = update_state(state, k3, dt)
    k4 = euler_step(tmp_state, control, vehicle_length)

    # Combine derivatives
    result = State()
    result.x = (k1.x + 2.0 * (k2.x + k3.x) + k4.x) / 6.0
    result.y = (k1.y + 2.0 * (k2.y + k3.y) + k4.y) / 6.0
    result.v = (k1.v + 2.0 * (k2.v + k3.v) + k4.v) / 6.0
    result.theta = (k1.theta + 2.0 * (k2.theta + k3.theta) + k4.theta) / 6.0

    return result


# Cost function implementations
@wp.func
def obstacle_cost(
    obstacles: wp.array(dtype=Obstacle),
    num_obstacles: wp.int32,
    px: wp.float32,
    py: wp.float32,
    vehicle_radius: wp.float32,
) -> wp.float32:
    """Compute collision cost with circular obstacles."""
    for i in range(num_obstacles):
        obstacle = obstacles[i]
        dx = obstacle.x - px
        dy = obstacle.y - py
        d_sq = dx * dx + dy * dy
        min_dist = obstacle.radius + vehicle_radius

        if d_sq < min_dist * min_dist:
            return 1e7  # Large penalty for collision

    return 0.0


@wp.func
def higgins_cost(
    obstacles: wp.array(dtype=Obstacle),
    num_obstacles: wp.int32,
    M: wp.float32,
    px: wp.float32,
    py: wp.float32,
    scan_range: wp.float32,
) -> wp.float32:
    """Higgins visibility cost function."""
    cost = 0.0
    r_fov_sq = scan_range * scan_range

    for i in range(num_obstacles):
        obstacle = obstacles[i]
        dx = obstacle.x - px
        dy = obstacle.y - py
        d_sq = dx * dx + dy * dy
        d = wp.sqrt(d_sq)

        inner = obstacle.radius / d * (r_fov_sq - d_sq)

        # Avoid overflow in exponential
        if inner > 60.0:
            score = inner
        else:
            score = wp.log(1.0 + wp.exp(inner))

        cost += M * score * score

    return cost


# Main MPPI kernel (stub)
@wp.kernel
def mppi_rollout_kernel(
    # Random states
    rand_states: wp.array(dtype=wp.uint32),
    # Trajectories
    x_nominal: wp.array(dtype=State),
    u_nominal: wp.array(dtype=Control),
    # Obstacles and costmap
    obstacles: wp.array(dtype=Obstacle),
    costmap: wp.array2d(dtype=wp.float32),
    # Parameters
    params: OptimizationParams,
    weights: CostWeights,
    limits: Limits,
    # Outputs
    u_disturbances: wp.array2d(dtype=Control),  # [samples, num_controls]
    costs: wp.array(dtype=wp.float32),  # [samples]
):
    """Main MPPI rollout kernel - computes costs for sampled trajectories."""
    sample_idx = wp.tid()

    if sample_idx >= params.samples:
        return

    # TODO: Implement full rollout logic
    # 1. Generate control disturbances
    # 2. Forward simulate trajectory
    # 3. Compute costs (state, control, obstacle, visibility)
    # 4. Store results

    # Placeholder - just store zero cost for now
    costs[sample_idx] = 0.0


class WarpMPPI:
    """
    Warp-based MPPI controller implementation.

    This class provides the same interface as the PyCUDA version but uses
    NVIDIA Warp for GPU acceleration.
    """

    def __init__(
        self,
        samples: int,
        vehicle_length: float,
        u_limits: List[float],
        u_dist_limits: List[float],
        Q: List[float],
        Qf: List[float],
        R: List[float],
        M: float = 1.0,
        c_lambda: float = 1.0,
        scan_range: float = 10.0,
        method: str = "Ours",
        debug: bool = False,
    ):
        """Initialize Warp-based MPPI controller."""

        self.samples = samples
        self.debug = debug

        # Create parameter structures
        self.params = OptimizationParams()
        self.params.samples = samples
        self.params.M = M
        self.params.vehicle_length = vehicle_length
        self.params.c_lambda = c_lambda
        self.params.scan_range = scan_range
        self.params.steering_rate_weight = 0.0
        self.params.method = 0  # TODO: Map method string to int

        self.weights = CostWeights()
        self.weights.Q = wp.vec4(Q[0], Q[1], Q[2], Q[3])
        self.weights.Qf = wp.vec4(Qf[0], Qf[1], Qf[2], Qf[3])
        self.weights.R = wp.vec2(R[0], R[1])

        self.limits = Limits()
        self.limits.u_limits = wp.vec2(u_limits[0], u_limits[1])
        self.limits.u_dist_limits = wp.vec2(u_dist_limits[0], u_dist_limits[1])

        if debug:
            print(f"[WarpMPPI] Initialized with {samples} samples")

    def find_control(
        self,
        costmap: np.ndarray,
        origin: Tuple[float, float],
        resolution: float,
        x_init: List[float],
        x_goal: List[float],
        x_nom: np.ndarray,
        u_nom: np.ndarray,
        obstacles: List[List[float]],
        dt: float,
    ) -> Tuple[np.ndarray, np.ndarray, np.ndarray]:
        """
        Find optimal control using Warp-accelerated MPPI.

        Args:
            costmap: Visibility/occupancy map
            origin: Map origin (x, y)
            resolution: Map resolution
            x_init: Initial state [x, y, v, theta]
            x_goal: Goal state [x, y, v, theta]
            x_nom: Nominal state trajectory
            u_nom: Nominal control sequence
            obstacles: List of obstacles [[x, y, radius], ...]
            dt: Time step

        Returns:
            Tuple of (optimal_controls, control_samples, weights)
        """

        # TODO: Implement Warp-based optimization
        # For now, return the nominal control as placeholder

        if self.debug:
            print(f"[WarpMPPI] Computing control with dt={dt:.3f}")

        # Placeholder implementation
        u_optimal = u_nom.copy()
        u_samples = np.zeros((self.samples, *u_nom.shape), dtype=np.float32)
        weights = np.ones(self.samples, dtype=np.float32) / self.samples

        return u_optimal, u_samples, weights


# Factory function for backward compatibility
def create_mppi_controller(**kwargs) -> WarpMPPI:
    """Create a Warp-based MPPI controller with the same interface."""
    return WarpMPPI(**kwargs)


if __name__ == "__main__":
    # Basic functionality test
    print("Testing Warp MPPI implementation...")

    # Create a simple controller
    controller = WarpMPPI(
        samples=1000,
        vehicle_length=2.5,
        u_limits=[3.0, 0.5],
        u_dist_limits=[0.5, 0.1],
        Q=[1.0, 1.0, 0.1, 0.1],
        Qf=[10.0, 10.0, 1.0, 1.0],
        R=[0.1, 0.1],
        debug=True,
    )

    print("✓ Warp MPPI controller created successfully!")
